Introducing control node setup script to ease the process for new users
This script will do the following for the user on the Ansible control node:
1. Register the node to CDN
2. Install epel
3. Install ansible
4. Install required packages for Ansible modules
5. git clone satellite-clone
6. Copies required config files in required locations

As explained above, this automates most of the manual steps in Ansible control node.  After running this script, the user just needs to plug in data in the created config files.

Support: The script supports RHEL6, RHEL7 and Fedora.

Fixes #55